feat: fill tenant id if development
